Send presampled events to Honeycomb (#8)
Ken does its own down-sampling logic, so events should be sent
"presampled" to Honeycomb. Previously, the Honeycomb client would _also_
apply the sample rate, resulting in unintentional double-sampling. This
means that a sample rate of 10 means you "see" 10% of the expected
events in Honeycomb's UI, which is part of why we never noticed -
however, that means you are only actually recording 1% of the events and
Honeycomb is scaling the aggregates back up by 10x.

To fix this, we should be using the `sendPresampled` event method, which
skips the `HoneyClient`'s internal downsampling logic.